return (int) instead of (size_t) from get_hight.

+++ 14-binary_tree_balance.c
@@ -6,9 +6,9 @@
  *
  * Return: the tree's height or 0 if tree is NULL.
  */
-size_t get_height(const binary_tree_t *tree)
+int get_height(const binary_tree_t *tree)
 {
-	size_t left = 0, right = 0;
+	int left = 0, right = 0;
 
 	if (tree == NULL)
 		return (0);
@@ -29,5 +29,5 @@ int binary_tree_balance(const binary_tree_t *tree)
 	if (tree == NULL)
 		return (0);
 
-	return ((int)(get_height(tree->left) - get_height(tree->right)));
+	return ((get_height(tree->left) - get_height(tree->right)));
 }
